Fixed height of elements in all types of roads
Must fix widht, so that they stick near the rumble.

+++ ModuleSceneStage.h
@@ -56,12 +56,12 @@ struct VisualElement {
 	SDL_Rect rect;
 	int x;
 	int y;
-	int worldPosition;
+	float worldPosition;
 	int nConsecutiveElements; 
 	VisualElementPosition position;
 
 	VisualElement() {};
-	VisualElement(SDL_Texture* texture, SDL_Rect rect, int x, int y, int worldPosition, int nConsecutiveElements, VisualElementPosition position) : 
+	VisualElement(SDL_Texture* texture, SDL_Rect rect, int x, int y, float worldPosition, int nConsecutiveElements, VisualElementPosition position) : 
 		texture(texture), rect(rect), x(x), y(y),	worldPosition(worldPosition), nConsecutiveElements(nConsecutiveElements), position(position) {};
 };
